# Conflicts:
#	smart-engine/src/main/java/org/smartdata/server/engine/StatesManager.java
#	smart-metastore/src/test/resources/druid-template.xml

+++ smart-metastore/src/main/java/org/smartdata/metastore/model/AggregatedAccessCounts.java
@@ -27,7 +27,7 @@ import org.smartdata.metrics.FileAccessEvent;
 @RequiredArgsConstructor
 public class AggregatedAccessCounts {
   private long fileId;
-  private final int accessCount;
+  private final long accessCount;
   private final long lastAccessedTimestamp;
 
   public AggregatedAccessCounts merge(AggregatedAccessCounts other) {
